Fix KOL list query type casting error
## Problem
KOL list API was failing with SQL type mismatch error:
- k.serial (kol_profiles) is VARCHAR (string)
- p.kol_serial (post_records) is INTEGER
- PostgreSQL couldn't join them without explicit type cast

## Solution
Added explicit type cast in LEFT JOIN:
- Changed: `ON k.serial = p.kol_serial`
- To: `ON k.serial::integer = p.kol_serial`

## Impact
- ✅ KOL list API now works correctly
- ✅ Returns proper statistics (total_posts, published_posts, interaction_rate)
- ✅ No more "operator does not exist" errors

+++ docker-container/finlab python/apps/unified-api/main.py
@@ -3797,7 +3797,7 @@ async def get_kol_list():
                         ELSE NULL
                     END), 0) as avg_interaction_rate
                 FROM kol_profiles k
-                LEFT JOIN post_records p ON k.serial = p.kol_serial
+                LEFT JOIN post_records p ON k.serial::integer = p.kol_serial
                 GROUP BY k.id, k.serial, k.nickname, k.member_id, k.persona, k.status, k.owner,
                          k.email, k.password, k.whitelist, k.notes, k.post_times, k.target_audience,
                          k.interaction_threshold, k.content_types, k.common_terms, k.colloquial_terms,
